requirements.txt
fastapi
uvicorn
gunicorn
ultralytics
starlette
requests
Pillow